Implement the q_reverse function
Add the function that can reverse the order of elements in a queue. The
detail implementation is that traverse all the nodes in the queue and
swap the value of the two pointers "next" and "prev" for each node.
Since the operation of swap pointers will change the order of queue,
we use the macro "list_for_each_safe" to traverse the nodes and use an
extra variable "prev" to keep the previous node for each. Finally,
swap the value of the two pointers of the head node and the reverse
of a queue is completed.