Implement q_delete_mid
Let two variable "next" and "prev" to traverse the queue from the head
node with "next" side and "prev" side respectively. Then, "next" node
will be the middle node if  "next" and "prev" point to the same node or
"prev" points to the the next node of "next" according to the number of
nodes in the queue. Finally use "list_del" to remove the node and free
the element that has the middle node. This implementation is still
uncompleted because there are some exception situation no implemented.