Implement q_delete_dup
This function is to delete the node whose string is duplicates to
the strings of other node. We traverse all nodes twice. The first
one is for recording the duplicate nodes while the second one is
for deleting the duplicate nodes. For the consideration of
performance, hash table is used in recording the nodes that appear
more than once or not.