Implement related functions for q_merge
This patch is to implement the function q_merge merging all sorted
queues to the first queue in the ascending order. There are two
parts inthe function: q_merge and merge_2_list. q_merge is to
traverse all queues in the chain except the first queue while
merge_2_list is to merge the queue that is visited  into the first
queue. And finally return the number of the elements in the first
queue after merged.