Implement the function q_descend
The function is to delete the least nodes in the queue and make the nodes
in the queue are in descending order. We traverse all nodes from the
"prev" direction and check the value of each node. We record the maximum
in the nodes that have been traversed and delete the node whose value is
less than maximum.