Modify the remove and free functions of queue
There are some problems in q_remove_head, q_remove_tail and q_free to
handle some issues of memory. In q_remove_head and q_remove_tail, use
strncpy for the copy of strings. And, in q_free, add the if-statement
to handle the situaion that the input is NULL.