Implement the new command "shuffle"
Implement shuffle command by Fisher-Yates shuffle algorithm. Since it's
convenient to implement in the form of array, we first declare an array
of pointer of type "list_head" to record the address of node in order.
Then, we shuffle the order in the array and finally add in the tail of
the head node to make those nodes in the form of doubly-linked list.