Modify the implementation in linenoise
To make the function linenoise also work in the situation that the
web server is activated, we try to use the function select() in
function line_edit() in order to watch both socket and stdin at the
same time. That linenoise function can still work in the command line
with web server activated. But, there are still some probelm that it
will show some message after typing command from web and need fixing.