Modify the hash table style in delete_dup
Introduce the Linux kernel style hash table in the delete_dup function. we use
the structure "hlist_node" and "hlist_head" which are declared in "linux/types.h"
and the APIs are descibed in "list.h" in the linux kernel headersd. Those declarations
and APIs are declared in the file "hlist.h" in this repository.With Linux kernel
style hash table, it is easier to know how the hash table works
in this function than the hash table from POSIX.